factoryAgent/pyproject.toml
[project]
name = "factoryagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-adk>=1.8.0",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
]
